Revert "Merge branch 'main' into v2.0-branch"

+++ app/src/display_module.h
@@ -106,7 +106,7 @@ void draw_scr_home(enum scroll_dir m_scroll_dir);
 void hpi_scr_home_update_spo2(int spo2);
 void hpi_scr_home_update_pr(int pr);
 void hpi_scr_home_update_rr(int rr);
-void hpi_scr_home_update_temp(double temp_f, double temp_c);
+void hpi_scr_home_update_temp(float temp_f, float temp_c);
 void hpi_scr_home_update_hr(int hr);
 void hpi_scr_home_update_lead_off(bool ecg_lead_off, bool ppg_lead_off);
